Refactor AddSomeGoals to tsx
Refactor AddSomeGoals to generic tsx
