+++ client/src/components/CreateNewGoal.jsx
@@ -1,6 +1,6 @@
 import {React, useState} from "react";
 import AddSomeSimple from "./addtolist/AddSomeSimple";
-import AddSomeGoals from "./addtolist/AddSomeGoals";
+import AddSomeGoals from "./addtolist/AddSomeGoals.tsx";
 import AddSomeBars from "./addtolist/AddSomeBars.tsx";
 
 export default function CreateNewGoal({tabs}) {
